fix some warnings that eeloo found
Some real issues here.

+++ IO/Xdmf2/Testing/Cxx/XdmfTestVTKIO.cxx
@@ -74,8 +74,14 @@ bool DoFilesExist(const char*xdmffile, const char*hdf5file, bool deleteIfSo)
   bool theyDo = xexists && xlenOK && hexists && hlenOK;
   if (theyDo && deleteIfSo && CleanUpGood)
   {
-    unlink(xdmffile);
-    unlink(hdf5file);
+    if (xdmffile)
+    {
+      unlink(xdmffile);
+    }
+    if (hdf5file)
+    {
+      unlink(hdf5file);
+    }
   }
 
   return theyDo;
